+++ packages/jest-worker/src/base/BaseWorkerPool.ts
@@ -6,6 +6,13 @@
  */
 
 import mergeStream = require('merge-stream');
+import {debug} from '../debug';
+
+//import type {MergedStream} from '@types/merge-stream'; // requires esModuleInterop
+interface MergedStream extends NodeJS.ReadWriteStream {
+    add(source: NodeJS.ReadableStream | ReadonlyArray<NodeJS.ReadableStream>): MergedStream;
+    isEmpty(): boolean;
+}
 import {
   CHILD_MESSAGE_END,
   PoolExitResult,
@@ -23,47 +30,85 @@ const FORCE_EXIT_DELAY = 500;
 const emptyMethod = () => {};
 
 export default class BaseWorkerPool {
-  private readonly _stderr: NodeJS.ReadableStream;
-  private readonly _stdout: NodeJS.ReadableStream;
+  private readonly _stderr: MergedStream;
+  private readonly _stdout: MergedStream;
   protected readonly _options: WorkerPoolOptions;
+  protected readonly _workerPath: string;
   private readonly _workers: Array<WorkerInterface>;
+  private readonly _workerTokens: Array<number | undefined>;
 
   constructor(workerPath: string, options: WorkerPoolOptions) {
+    this._workerPath = workerPath;
     this._options = options;
-    this._workers = new Array(options.numWorkers);
 
-    const stdout = mergeStream();
-    const stderr = mergeStream();
+    this._stdout = mergeStream();
+    this._stderr = mergeStream();
 
-    const {forkOptions, maxRetries, resourceLimits, setupArgs} = options;
-
-    for (let i = 0; i < options.numWorkers; i++) {
-      const workerOptions: WorkerOptions = {
-        forkOptions,
-        maxRetries,
-        resourceLimits,
-        setupArgs,
-        workerId: i,
-        workerPath,
-      };
-
-      const worker = this.createWorker(workerOptions);
-      const workerStdout = worker.getStdout();
-      const workerStderr = worker.getStderr();
-
-      if (workerStdout) {
-        stdout.add(workerStdout);
-      }
-
-      if (workerStderr) {
-        stderr.add(workerStderr);
-      }
-
-      this._workers[i] = worker;
+    if (options.jobClient) {
+      // with jobClient, override numWorkers
+      // NOTE maxJobs is upper bound. minJobs is 1
+      options.numWorkers = options.jobClient.maxJobs;
     }
 
-    this._stdout = stdout;
-    this._stderr = stderr;
+    this._workers = new Array(options.numWorkers);
+    this._workerTokens = new Array(options.numWorkers);
+
+    if (options.jobClient) {
+      // start 1 worker, grow on demand
+      this._addWorker(0, true);
+    }
+    else {
+      debug(`BaseWorkerPool.constructor: no jobClient -> start all workers before demand`)
+      // no jobclient -> numWorkers is static
+      // start all workers before demand
+      for (let i = 0; i < options.numWorkers; i++) {
+        this._addWorker(i);
+      }
+    }
+  }
+
+  _addWorker(workerId: number, ignoreJobClient = false): WorkerInterface | null {
+    debug(`BaseWorkerPool._addWorker`)
+    // ignoreJobClient is true for the first worker with jobclient
+
+    if (this._workers[workerId]) return this._workers[workerId];
+
+    if (this._options.jobClient && !ignoreJobClient) {
+      const numWorkers = this._workers.filter(Boolean).length;
+      if (numWorkers > 0) {
+        const token = this._options.jobClient.acquire();
+        if (token == null) {
+          return null; // jobserver is full, try again later
+        }
+        this._workerTokens[workerId] = token;
+      }
+      // else: dont acquire token for the first worker
+    }
+
+    const {forkOptions, maxRetries, resourceLimits, setupArgs} = this._options;
+    const workerOptions: WorkerOptions = {
+      forkOptions,
+      maxRetries,
+      resourceLimits,
+      setupArgs,
+      workerId,
+      workerPath: this._workerPath,
+    };
+
+    const worker = this.createWorker(workerOptions);
+    const workerStdout = worker.getStdout();
+    const workerStderr = worker.getStderr();
+
+    if (workerStdout) {
+      this._stdout.add(workerStdout);
+    }
+
+    if (workerStderr) {
+      this._stderr.add(workerStderr);
+    }
+
+    this._workers[workerId] = worker;
+    return worker;
   }
 
   getStderr(): NodeJS.ReadableStream {
@@ -78,7 +123,12 @@ export default class BaseWorkerPool {
     return this._workers;
   }
 
-  getWorkerById(workerId: number): WorkerInterface {
+  getWorkerById(workerId: number): WorkerInterface | null {
+    if (this._options.jobClient && this._workers[workerId] == undefined) {
+      // try to create new worker
+      debug(`BaseWorkerPool.getWorkerById: create worker ${workerId}`);
+      return this._addWorker(workerId);
+    }
     return this._workers[workerId];
   }
 
@@ -89,7 +139,8 @@ export default class BaseWorkerPool {
   async end(): Promise<PoolExitResult> {
     // We do not cache the request object here. If so, it would only be only
     // processed by one of the workers, and we want them all to close.
-    const workerExitPromises = this._workers.map(async worker => {
+    const workerExitPromises = this._workers.map(async (worker, workerId) => {
+      if (!worker) return false;
       worker.send(
         [CHILD_MESSAGE_END, false],
         emptyMethod,
@@ -108,6 +159,14 @@ export default class BaseWorkerPool {
       await worker.waitForExit();
       // Worker ideally exited gracefully, don't send force exit then
       clearTimeout(forceExitTimeout);
+
+      if (this._options.jobClient) {
+        const token = this._workerTokens[workerId];
+        if (token != undefined) {
+          this._options.jobClient.release(token);
+          this._workerTokens[workerId] = undefined;
+        }
+      }
 
       return forceExited;
     });